chore: Migrate CI to forgejo
The development scripts and related Kubernetes resources were updated to
reflect the change from Gitea to Forgejo. This involved renaming files,
updating variable names, and modifying configuration to use Forgejo's
specific paths and image. The core functionality of setting up a Git
provider for testing purposes remains the same.

Add some comment on policy.go so the e2e can trigger and test this
change.

Jira: https://issues.redhat.com/browse/SRVKP-9991
AI-assisted-by: Google Gemini
