+++ hack/dev/kind/README.md
@@ -1,3 +1,3 @@
 # Usage
 
-See <https://pipelinesascode.com/docs/dev/> for how to use it.
+Use [startpaac](https://github.com/chmouel/startpaac/) which is better for local dev or if you really want to use those script you can go over <https://pipelinesascode.com/docs/dev/> for how to use it.
