+++ test/pkg/gitea/test.go
@@ -130,7 +130,7 @@ func TestPR(t *testing.T, topts *TestOpts) (context.Context, func()) {
 	hookURL := os.Getenv("TEST_GITEA_SMEEURL")
 	topts.InternalGiteaURL = os.Getenv("TEST_GITEA_INTERNAL_URL")
 	if topts.InternalGiteaURL == "" {
-		topts.InternalGiteaURL = "http://gitea.gitea:3000"
+		topts.InternalGiteaURL = "http://forgejo.forgejo:3000"
 	}
 	if topts.ExtraArgs == nil {
 		topts.ExtraArgs = map[string]string{}
@@ -301,7 +301,7 @@ func NewPR(t *testing.T, topts *TestOpts) func() {
 	topts.GiteaPassword = giteaPassword
 	topts.InternalGiteaURL = os.Getenv("TEST_GITEA_INTERNAL_URL")
 	if topts.InternalGiteaURL == "" {
-		topts.InternalGiteaURL = "http://gitea.gitea:3000"
+		topts.InternalGiteaURL = "http://forgejo.forgejo:3000"
 	}
 	if topts.ExtraArgs == nil {
 		topts.ExtraArgs = map[string]string{}
